refactor(proxy-gateway): move the Responses payload guards out of the service
Group 3 continued. Behavior preserving.

Four shape guards that narrow loosely typed upstream payloads into the Responses
surface types now live in `modules/openai/responses/openai-responses-adapters.ts`.
Like the conversion functions in the previous commit they only referenced each
other, never instance state, so this is a move rather than a redesign.

Nothing outside the service reaches them, by name or reflectively, so no
delegates were needed this time. That was checked rather than assumed: the
previous split in this series turned out to have two call sites reaching in
through `Reflect.get`.

  openai.service.ts   1007 -> 888 lines  (1394 before this group)
  responses/openai-responses-adapters.ts   new, 135 lines

Verified: `tsc --noEmit` clean, `eslint .` 0 errors with the same 8 warnings as
the base, `prettier --check .` clean, `vitest run` 818 passed and 4 skipped with
no failures.
